[Silver V] Title: 타일 장식물, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13301. 타일 장식물/타일 장식물.cc
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-    int n,n1=4,n2=6,n3;
+    long long int n,n1=4,n2=6,n3;
     cin >> n;
     if (n>=3){
         for (int i=3;i<=n;i++){
